♻️ refactor: EditProfileModal 컴포넌트 리팩토링
뷰와 로직 분리
